056 ROLES - Migracion tabla roles

+++ database/migrations/2020_06_18_204014_create_proveedores_table.php
@@ -14,7 +14,7 @@ class CreateProveedoresTable extends Migration
     public function up()
     {
         Schema::create('proveedores', function (Blueprint $table) {
-            $table->integer('id')->unsigned();
+            $table->bigInteger('id')->unsigned();
             $table->string('contacto',50)->nullable();
             $table->string('telefono_contacto',50)->nullable();
             $table->foreign('id')->references('id')->on('personas')->onDelete('cascade');
